refactor: file and code blocks restructuring
