+++ india_compliance/gst_india/doctype/gstr_1_beta/gstr_1_beta.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2024, Resilient Tech and contributors
 # For license information, please see license.txt
 
-import json
 from datetime import datetime
 
 import frappe
@@ -11,45 +10,17 @@ from frappe.model.document import Document
 from frappe.query_builder.functions import Date, Sum
 from frappe.utils import flt, get_last_day, getdate
 
-from india_compliance.gst_india.doctype.gstr_1_filed_log.gstr_1_filed_log import (
-    summarize_data,
-)
-from india_compliance.gst_india.report.gstr_1.gstr_1 import (
-    GSTR1DocumentIssuedSummary,
-    GSTR11A11BData,
-)
-from india_compliance.gst_india.utils.__init__ import get_gst_accounts_by_type
-from india_compliance.gst_india.utils.exporter import ExcelExporter
+from india_compliance.gst_india.utils import get_gst_accounts_by_type
 from india_compliance.gst_india.utils.gstin_info import get_gstr_1_return_status
-from india_compliance.gst_india.utils.gstr_1 import (
-    ExcelWidth,
-    GSTR1_Categories,
-    GSTR1_DataFields,
-    GSTR1_Excel_Categories,
-    GSTR1_ItemFields,
-    GSTR1_SubCategories,
-    get_file_name,
-)
-from india_compliance.gst_india.utils.gstr_1.gstr_1_data import GSTR1Invoices
+from india_compliance.gst_india.utils.gstr_1 import GSTR1_SubCategories
 from india_compliance.gst_india.utils.gstr_1.gstr_1_download import (
     download_gstr1_json_data,
 )
 from india_compliance.gst_india.utils.gstr_1.gstr_1_json_map import (
-    convert_to_gov_data_format,
+    GSTR1BooksData,
     summarize_retsum_data,
 )
 from india_compliance.gst_india.utils.gstr_utils import request_otp
-
-AMOUNT_DATA_FORMAT = "#,##0.00"
-DATE_FORMAT = "dd-mmm-yy"
-
-AMOUNT_FIELDS = {
-    GSTR1_DataFields.TAXABLE_VALUE.value: 0,
-    GSTR1_DataFields.IGST.value: 0,
-    GSTR1_DataFields.CGST.value: 0,
-    GSTR1_DataFields.SGST.value: 0,
-    GSTR1_DataFields.CESS.value: 0,
-}
 
 
 class GSTR1Beta(Document):
@@ -148,7 +119,7 @@ class GSTR1Beta(Document):
 
         # generate gstr1
         gstr1_log.update_status("In Progress")
-        frappe.enqueue(self.generate_gstr1, queue="short", now=True)
+        frappe.enqueue(self.generate_gstr1, queue="short")
         frappe.msgprint("GSTR-1 is being prepared", alert=True)
 
     def generate_gstr1(self):
@@ -188,7 +159,7 @@ class GSTR1Beta(Document):
                 doctype=self.doctype,
             )
 
-        def _summarize_data(gov_data_field=None):
+        def summarize_data(gov_data_field=None):
             summary_fields = {
                 "reconcile": "reconcile_summary",
                 f"{gov_data_field}": f"{gov_data_field}_summary",
@@ -206,19 +177,19 @@ class GSTR1Beta(Document):
                 if key == "filed":
                     summary_data = summarize_retsum_data(data[key].get("summary"))
                 else:
-                    summary_data = summarize_data(data[key])
+                    summary_data = self.gstr1_log.summarize_data(data[key])
 
                 self.gstr1_log.update_json_for(field, summary_data)
                 data[field] = summary_data
 
         # APIs Disabled
         if not self.gstr1_log.is_gstr1_api_enabled():
-            books_data = compute_books_gstr1_data(self)
+            books_data = get_books_gstr1_data(self)
 
             data["status"] = "Not Filed"
             data["books"] = self.gstr1_log.normalize_data(books_data)
 
-            _summarize_data()
+            summarize_data()
             on_generate()
             return
 
@@ -238,7 +209,7 @@ class GSTR1Beta(Document):
             gov_data_field = "unfiled"
 
         # Get Data
-        gov_data, is_enqueued = get_gstr1_json_data(self.gstr1_log)
+        gov_data, is_enqueued = get_gov_gstr1_data(self.gstr1_log)
 
         if error_type := gov_data.get("error_type"):
             # otp_requested, invalid_otp
@@ -250,12 +221,12 @@ class GSTR1Beta(Document):
             on_generate()
             return
 
-        books_data = compute_books_gstr1_data(self)
+        books_data = get_books_gstr1_data(self)
 
         if is_enqueued:
             return
 
-        reconcile_data = reconcile_gstr1_data(self.gstr1_log, gov_data, books_data)
+        reconcile_data = get_reconcile_gstr1_data(self.gstr1_log, gov_data, books_data)
 
         # Compile Data
         data["status"] = status
@@ -264,28 +235,14 @@ class GSTR1Beta(Document):
         data[gov_data_field] = self.gstr1_log.normalize_data(gov_data)
         data["books"] = self.gstr1_log.normalize_data(books_data)
 
-        _summarize_data(gov_data_field)
+        summarize_data(gov_data_field)
         on_generate()
 
 
-def get_period(month_or_quarter, year):
-    if "-" in month_or_quarter:
-        # Quarterly
-        last_month = month_or_quarter.split("-")[1]
-        month_number = str(getdate(f"{last_month}-{year}").month).zfill(2)
-
-    else:
-        # Monthly
-        month_number = str(datetime.strptime(month_or_quarter, "%B").month).zfill(2)
-
-    return f"{month_number}{year}"
+####### DATA ######################################################################################
 
 
-def generate_gstr1():
-    pass
-
-
-def get_gstr1_json_data(gstr1_log):
+def get_gov_gstr1_data(gstr1_log):
     if gstr1_log.filing_status == "Filed":
         data_field = "filed"
 
@@ -302,7 +259,7 @@ def get_gstr1_json_data(gstr1_log):
     return download_gstr1_json_data(gstr1_log)
 
 
-def compute_books_gstr1_data(filters, save=False, periodicity="Monthly"):
+def get_books_gstr1_data(filters):
     # Query / Process / Map / Sumarize / Optionally Save & Return
     data_field = "books"
     gstr1_log = filters.gstr1_log
@@ -326,28 +283,14 @@ def compute_books_gstr1_data(filters, save=False, periodicity="Monthly"):
     )
 
     # compute data
-    mapped_data = GSTR1MappedData(_filters).prepare_mapped_data()
+    mapped_data = GSTR1BooksData(_filters).prepare_mapped_data()
 
     gstr1_log.update_json_for(data_field, mapped_data)
 
     return mapped_data
 
 
-def get_from_and_to_date(month_or_quarter, year):
-    filing_frequency = get_gstr1_filing_frequency()
-
-    if filing_frequency == "Monthly":
-        from_date = getdate(f"{year}-{month_or_quarter}-01")
-        to_date = get_last_day(from_date)
-    else:
-        start_month, end_month = month_or_quarter.split("-")
-        from_date = getdate(f"{year}-{start_month}-01")
-        to_date = get_last_day(f"{year}-{end_month}-01")
-
-    return from_date, to_date
-
-
-def reconcile_gstr1_data(gstr1_log, gov_data, books_data):
+def get_reconcile_gstr1_data(gstr1_log, gov_data, books_data):
     # Everything from gov_data compared with books_data
     # Missing in gov_data
     # Update books data (optionally if not filed)
@@ -373,7 +316,7 @@ def reconcile_gstr1_data(gstr1_log, gov_data, books_data):
             GSTR1_SubCategories.HSN.value,
             GSTR1_SubCategories.DOC_ISSUE.value,
         ]
-        is_list = False
+        is_list = False  # Object Type for the subdata_value
 
         reconcile_subdata = {}
 
@@ -513,53 +456,14 @@ def get_reconciled_row(books_row, gov_row):
     return reconcile_row
 
 
-def get_empty_row(row: dict):
-    empty_row = row.copy()
-
-    for key, value in empty_row.items():
-        if isinstance(value, (int, float)):
-            empty_row[key] = 0
-
-    return empty_row
-
-
-def get_aggregated_row(books_rows: list) -> dict:
-    """
-    There can be multiple rows in books data for a single row in gov data
-    Aggregate all the rows to a single row
-    """
-    aggregated_row = {}
-
-    for row in books_rows:
-        if not aggregated_row:
-            aggregated_row = row.copy()
-            continue
-
-        for key, value in row.items():
-            if isinstance(value, (int, float)):
-                aggregated_row[key] += value
-
-    return aggregated_row
-
-
-###################
-def get_gstr1_filing_frequency():
-    gst_settings = frappe.get_cached_doc("GST Settings")
-    return gst_settings.filing_frequency
-
-
-@frappe.whitelist()
-def is_latest_data(company_gstin, month_or_quarter, year):
-    period = get_period(month_or_quarter, year)
-    if log_name := frappe.db.exists("GSTR-1 Filed Log", f"{period}-{company_gstin}"):
-        gstr1_log = frappe.get_doc("GSTR-1 Filed Log", log_name)
-        return gstr1_log.is_latest_data
-
-    return True
-
-
 @frappe.whitelist()
 def get_output_gst_balance(company, company_gstin, month_or_quarter, year):
+    """
+    Returns the net output balance for the given return period
+    """
+
+    frappe.has_permission("GSTR-1 Beta", throw=True)
+
     from_date, to_date = get_from_and_to_date(month_or_quarter, year)
 
     filters = frappe._dict(
@@ -595,3437 +499,78 @@ def get_output_gst_balance(company, company_gstin, month_or_quarter, year):
     return net_output_balance
 
 
-####################################################################################################
-####### DOWNLOAD APIs ##############################################################################
-####################################################################################################
+####### UTILS ######################################################################################
 
 
-@frappe.whitelist()
-def download_books_as_excel(company_gstin, month_or_quarter, year):
+def get_period(month_or_quarter: str, year: str) -> str:
+    """
+    Returns the period in the format MMYYYY
+    as accepted by the GST Portal
+    """
 
-    books_excel = BooksExcel(company_gstin, month_or_quarter, year)
-    books_excel.export_data()
+    if "-" in month_or_quarter:
+        # Quarterly
+        last_month = month_or_quarter.split("-")[1]
+        month_number = str(getdate(f"{last_month}-{year}").month).zfill(2)
 
-    return "Data Downloaded to Excel Successfully"
+    else:
+        # Monthly
+        month_number = str(datetime.strptime(month_or_quarter, "%B").month).zfill(2)
 
-
-@frappe.whitelist()
-def download_reconcile_as_excel(company_gstin, month_or_quarter, year):
-
-    reconcile_excel = ReconcileExcel(company_gstin, month_or_quarter, year)
-    reconcile_excel.export_data()
-
-    return "Data Downloaded to Excel Successfully"
+    return f"{month_number}{year}"
 
 
-@frappe.whitelist()
-def download_gstr_1_json(
-    company_gstin,
-    year,
-    month_or_quarter,
-    include_uploaded=False,
-    delete_missing=False,
-):
-    if isinstance(include_uploaded, str):
-        include_uploaded = json.loads(include_uploaded)
+def get_from_and_to_date(month_or_quarter: str, year: str) -> tuple:
+    """
+    Returns the from and to date for the given month or quarter and year
+    This is used to filter the data for the given period in Books
+    """
 
-    if isinstance(delete_missing, str):
-        delete_missing = json.loads(delete_missing)
+    filing_frequency = get_gstr1_filing_frequency()
 
-    period = get_period(month_or_quarter, year)
-    gstr1_log = frappe.get_doc("GSTR-1 Filed Log", f"{period}-{company_gstin}")
+    if filing_frequency == "Monthly":
+        from_date = getdate(f"{year}-{month_or_quarter}-01")
+        to_date = get_last_day(from_date)
+    else:
+        start_month, end_month = month_or_quarter.split("-")
+        from_date = getdate(f"{year}-{start_month}-01")
+        to_date = get_last_day(f"{year}-{end_month}-01")
 
-    data = gstr1_log.get_json_for("books")
+    return from_date, to_date
 
-    for subcategory_data in data.values():
-        discard_invoices = []
 
-        if isinstance(subcategory_data, str):
+def get_gstr1_filing_frequency():
+    gst_settings = frappe.get_cached_doc("GST Settings")
+    return gst_settings.filing_frequency
+
+
+def get_empty_row(row: dict):
+    """
+    Row with all values as 0
+    """
+    empty_row = row.copy()
+
+    for key, value in empty_row.items():
+        if isinstance(value, (int, float)):
+            empty_row[key] = 0
+
+    return empty_row
+
+
+def get_aggregated_row(books_rows: list) -> dict:
+    """
+    There can be multiple rows in books data for a single row in gov data
+    Aggregate all the rows to a single row
+    """
+    aggregated_row = {}
+
+    for row in books_rows:
+        if not aggregated_row:
+            aggregated_row = row.copy()
             continue
 
-        for key, row in subcategory_data.items():
-            if isinstance(row, list):
-                row = row[0]
-
-            if not row.get("upload_status"):
-                continue
-
-            if row.get("upload_status") == "Uploaded" and not include_uploaded:
-                discard_invoices.append(key)
-
-            if row.get("upload_status") == "Missing in Books":
-                if delete_missing:
-                    row["flag"] = "D"
-                else:
-                    discard_invoices.append(key)
-
-        for key in discard_invoices:
-            subcategory_data.pop(key)
-
-    return {
-        "data": {
-            "version": "GST3.0.4",
-            "gstin": company_gstin,
-            "hash": "hash",
-            "fp": period,
-            **convert_to_gov_data_format(data),
-        },
-        "filename": f"GSTR-1-gov-{company_gstin}-{period}.json",
-    }
-
-
-#################################
-##### Process Data #############
-################################
-class GSTR1ProcessData:
-
-    def get_transaction_type(self, invoice):
-        if invoice.is_debit_note:
-            return "Debit Note"
-        elif invoice.is_return:
-            return "Credit Note"
-        else:
-            return "Invoice"
-
-    def process_data_for_invoice_no_key(self, invoice, prepared_data):
-        invoice_sub_category = invoice.invoice_sub_category
-        invoice_no = invoice.invoice_no
-
-        mapped_dict = prepared_data.setdefault(invoice_sub_category, {}).setdefault(
-            invoice_no,
-            {
-                GSTR1_DataFields.TRANSACTION_TYPE.value: self.get_transaction_type(
-                    invoice
-                ),
-                GSTR1_DataFields.CUST_GSTIN.value: invoice.billing_address_gstin,
-                GSTR1_DataFields.CUST_NAME.value: invoice.customer_name,
-                GSTR1_DataFields.DOC_DATE.value: invoice.posting_date,
-                GSTR1_DataFields.DOC_NUMBER.value: invoice.invoice_no,
-                GSTR1_DataFields.DOC_VALUE.value: invoice.invoice_total,
-                GSTR1_DataFields.POS.value: invoice.place_of_supply,
-                GSTR1_DataFields.REVERSE_CHARGE.value: (
-                    "Y" if invoice.is_reverse_charge else "N"
-                ),
-                GSTR1_DataFields.DOC_TYPE.value: invoice.invoice_type,
-                GSTR1_DataFields.TAXABLE_VALUE.value: 0,
-                GSTR1_DataFields.IGST.value: 0,
-                GSTR1_DataFields.CGST.value: 0,
-                GSTR1_DataFields.SGST.value: 0,
-                GSTR1_DataFields.CESS.value: 0,
-                GSTR1_DataFields.DIFF_PERCENTAGE.value: 0,
-                "items": [],
-            },
-        )
-
-        items = mapped_dict["items"]
-
-        for item in items:
-            if item[GSTR1_ItemFields.TAX_RATE.value] == invoice.gst_rate:
-                item[GSTR1_ItemFields.TAXABLE_VALUE.value] += invoice.taxable_value
-                item[GSTR1_ItemFields.IGST.value] += invoice.igst_amount
-                item[GSTR1_ItemFields.CGST.value] += invoice.cgst_amount
-                item[GSTR1_ItemFields.SGST.value] += invoice.sgst_amount
-                item[GSTR1_ItemFields.CESS.value] += invoice.total_cess_amount
-                self.update_totals(mapped_dict, invoice)
-                return
-
-        items.append(
-            {
-                GSTR1_ItemFields.TAXABLE_VALUE.value: invoice.taxable_value,
-                GSTR1_ItemFields.IGST.value: invoice.igst_amount,
-                GSTR1_ItemFields.CGST.value: invoice.cgst_amount,
-                GSTR1_ItemFields.SGST.value: invoice.sgst_amount,
-                GSTR1_ItemFields.CESS.value: invoice.total_cess_amount,
-                GSTR1_ItemFields.TAX_RATE.value: invoice.gst_rate,
-            }
-        )
-
-        self.update_totals(mapped_dict, invoice)
-
-    def process_data_for_document_category_key(self, invoice, prepared_data):
-        key = invoice.invoice_category
-        mapped_dict = prepared_data.setdefault(key, {}).setdefault(
-            invoice.invoice_type, []
-        )
-
-        for row in mapped_dict:
-            if row[GSTR1_DataFields.DOC_NUMBER.value] == invoice.invoice_no:
-                self.update_totals(row, invoice)
-                return
-
-        mapped_dict.append(
-            {
-                GSTR1_DataFields.TRANSACTION_TYPE.value: self.get_transaction_type(
-                    invoice
-                ),
-                GSTR1_DataFields.DOC_NUMBER.value: invoice.invoice_no,
-                GSTR1_DataFields.DOC_DATE.value: invoice.posting_date,
-                **self.get_invoice_values(invoice),
-            }
-        )
-
-    def process_data_for_b2cs(self, invoice, prepared_data):
-        key = f"{invoice.place_of_supply} - {flt(invoice.gst_rate)}"
-        mapped_dict = prepared_data.setdefault("B2C (Others)", {}).setdefault(key, [])
-
-        for row in mapped_dict:
-            if row[GSTR1_DataFields.DOC_NUMBER.value] == invoice.invoice_no:
-                self.update_totals(row, invoice)
-                return
-
-        mapped_dict.append(
-            {
-                GSTR1_DataFields.DOC_DATE.value: invoice.posting_date,
-                GSTR1_DataFields.DOC_NUMBER.value: invoice.invoice_no,
-                GSTR1_DataFields.CUST_NAME.value: invoice.customer_name,
-                # currently other value is not supported in GSTR-1
-                GSTR1_DataFields.DOC_TYPE.value: "OE",
-                GSTR1_DataFields.TRANSACTION_TYPE.value: self.get_transaction_type(
-                    invoice
-                ),
-                GSTR1_DataFields.POS.value: invoice.place_of_supply,
-                GSTR1_DataFields.TAX_RATE.value: invoice.gst_rate,
-                GSTR1_DataFields.ECOMMERCE_GSTIN.value: invoice.ecommerce_gstin,
-                **self.get_invoice_values(invoice),
-            }
-        )
-
-    def process_data_for_hsn_summary(self, invoice, prepared_data):
-        key = f"{invoice.gst_hsn_code} - {invoice.stock_uom} - {flt(invoice.gst_rate)}"
-
-        if key not in prepared_data:
-            mapped_dict = prepared_data.setdefault(
-                key,
-                {
-                    GSTR1_DataFields.HSN_CODE.value: invoice.gst_hsn_code,
-                    GSTR1_DataFields.DESCRIPTION.value: frappe.db.get_value(
-                        "GST HSN Code", invoice.gst_hsn_code, "description"
-                    ),
-                    GSTR1_DataFields.UOM.value: invoice.stock_uom,
-                    GSTR1_DataFields.QUANTITY.value: 0,
-                    GSTR1_DataFields.TAX_RATE.value: invoice.gst_rate,
-                    GSTR1_DataFields.TAXABLE_VALUE.value: 0,
-                    GSTR1_DataFields.IGST.value: 0,
-                    GSTR1_DataFields.CGST.value: 0,
-                    GSTR1_DataFields.SGST.value: 0,
-                    GSTR1_DataFields.CESS.value: 0,
-                },
-            )
-
-        else:
-            mapped_dict = prepared_data[key]
-
-        self.update_totals(mapped_dict, invoice, for_qty=True)
-
-    def process_data_for_document_issued_summary(self, row, prepared_data):
-        key = f"{row['nature_of_document']} - {row['from_serial_no']}"
-        prepared_data.setdefault(
-            key,
-            {
-                GSTR1_DataFields.DOC_TYPE.value: row["nature_of_document"],
-                GSTR1_DataFields.FROM_SR.value: row["from_serial_no"],
-                GSTR1_DataFields.TO_SR.value: row["to_serial_no"],
-                GSTR1_DataFields.TOTAL_COUNT.value: row["total_issued"],
-                GSTR1_DataFields.DRAFT_COUNT.value: row["total_draft"],
-                GSTR1_DataFields.CANCELLED_COUNT.value: row["cancelled"],
-            },
-        )
-
-    def process_data_for_advances_received_or_adjusted(self, row, prepared_data):
-        advances = {}
-        tax_rate = round(((row["tax_amount"] / row["taxable_value"]) * 100))
-        key = f"{row['place_of_supply']} - {flt(tax_rate)}"
-
-        mapped_dict = prepared_data.setdefault(key, [])
-
-        advances[GSTR1_DataFields.CUST_NAME.value] = row["party"]
-        advances[GSTR1_DataFields.DOC_NUMBER.value] = row["name"]
-        advances[GSTR1_DataFields.DOC_DATE.value] = row["posting_date"]
-        advances[GSTR1_DataFields.POS.value] = row["place_of_supply"]
-        advances[GSTR1_DataFields.TAXABLE_VALUE.value] = row["taxable_value"]
-        advances[GSTR1_DataFields.TAX_RATE.value] = tax_rate
-        advances[GSTR1_DataFields.CESS.value] = row["cess_amount"]
-
-        if row.get("reference_name"):
-            advances["against_voucher"] = row["reference_name"]
-
-        if row["place_of_supply"][0:2] == row["company_gstin"][0:2]:
-            advances[GSTR1_DataFields.CGST.value] = row["tax_amount"] / 2
-            advances[GSTR1_DataFields.SGST.value] = row["tax_amount"] / 2
-            advances[GSTR1_DataFields.IGST.value] = 0
-
-        else:
-            advances[GSTR1_DataFields.IGST.value] = row["tax_amount"]
-            advances[GSTR1_DataFields.CGST.value] = 0
-            advances[GSTR1_DataFields.SGST.value] = 0
-
-        mapped_dict.append(advances)
-
-    # utils
-
-    def update_totals(self, mapped_dict, invoice, for_qty=False):
-        data_invoice_amount_map = {
-            GSTR1_DataFields.TAXABLE_VALUE.value: GSTR1_ItemFields.TAXABLE_VALUE.value,
-            GSTR1_DataFields.IGST.value: GSTR1_ItemFields.IGST.value,
-            GSTR1_DataFields.CGST.value: GSTR1_ItemFields.CGST.value,
-            GSTR1_DataFields.SGST.value: GSTR1_ItemFields.SGST.value,
-            GSTR1_DataFields.CESS.value: GSTR1_ItemFields.CESS.value,
-        }
-
-        if for_qty:
-            data_invoice_amount_map[GSTR1_DataFields.QUANTITY.value] = "qty"
-
-        for key, field in data_invoice_amount_map.items():
-            mapped_dict[key] += invoice.get(field, 0)
-
-    def get_invoice_values(self, invoice):
-        return {
-            GSTR1_DataFields.TAXABLE_VALUE.value: invoice.taxable_value,
-            GSTR1_DataFields.IGST.value: invoice.igst_amount,
-            GSTR1_DataFields.CGST.value: invoice.cgst_amount,
-            GSTR1_DataFields.SGST.value: invoice.sgst_amount,
-            GSTR1_DataFields.CESS.value: invoice.total_cess_amount,
-        }
-
-
-def test_gstr1_mapped_data():
-    filters = frappe._dict(
-        {
-            "company": "Shalibhadra Metal Corporation",
-            "company_gstin": "24AAUPV7468F1ZW",
-            "from_date": getdate("2024-03-01"),
-            "to_date": get_last_day("2024-03-01"),
-        }
-    )
-
-    return GSTR1MappedData(filters).prepare_mapped_data()
-
-
-class GSTR1MappedData(GSTR1ProcessData):
-    def __init__(self, filters):
-        self.filters = filters
-
-    def prepare_mapped_data(self):
-        prepared_data = {}
-
-        _class = GSTR1Invoices(self.filters)
-        data = _class.get_invoices_for_item_wise_summary()
-        _class.process_invoices(data)
-
-        for invoice in data:
-            if invoice["invoice_category"] in (
-                GSTR1_Categories.B2B.value,
-                GSTR1_Categories.EXP.value,
-                GSTR1_Categories.B2CL.value,
-                GSTR1_Categories.CDNR.value,
-                GSTR1_Categories.CDNUR.value,
-            ):
-                self.process_data_for_invoice_no_key(invoice, prepared_data)
-            elif invoice["invoice_category"] == GSTR1_Categories.NIL_EXEMPT.value:
-                self.process_data_for_document_category_key(invoice, prepared_data)
-            elif invoice["invoice_category"] == GSTR1_Categories.B2CS.value:
-                self.process_data_for_b2cs(invoice, prepared_data)
-
-        other_categories = {
-            GSTR1_Categories.AT.value: self.prepare_advances_recevied_data(),
-            GSTR1_Categories.TXP.value: self.prepare_advances_adjusted_data(),
-            GSTR1_Categories.HSN.value: self.prepare_hsn_data(data),
-            GSTR1_Categories.DOC_ISSUE.value: self.prepare_document_issued_data(),
-        }
-
-        for category, data in other_categories.items():
-            if data:
-                prepared_data[category] = data
-
-        return prepared_data
-
-    def prepare_document_issued_data(self):
-        doc_issued_data = {}
-        data = GSTR1DocumentIssuedSummary(self.filters).get_data()
-
-        for row in data:
-            self.process_data_for_document_issued_summary(row, doc_issued_data)
-
-        return doc_issued_data
-
-    def prepare_hsn_data(self, data):
-        hsn_summary_data = {}
-
-        for row in data:
-            self.process_data_for_hsn_summary(row, hsn_summary_data)
-
-        return hsn_summary_data
-
-    def prepare_advances_recevied_data(self):
-        return self.prepare_advances_received_or_adjusted_data("Advances")
-
-    def prepare_advances_adjusted_data(self):
-        return self.prepare_advances_received_or_adjusted_data("Adjustment")
-
-    def prepare_advances_received_or_adjusted_data(self, type_of_business):
-        advances_data = {}
-        self.filters.type_of_business = type_of_business
-        gst_accounts = get_gst_accounts_by_type(self.filters.company, "Output")
-        _class = GSTR11A11BData(self.filters, gst_accounts)
-
-        if type_of_business == "Advances":
-            query = _class.get_11A_query()
-            fields = (
-                _class.pe.name,
-                _class.pe.party,
-                _class.pe.posting_date,
-                _class.pe.company_gstin,
-            )
-        elif type_of_business == "Adjustment":
-            query = _class.get_11B_query()
-            fields = (
-                _class.pe.name,
-                _class.pe.party,
-                _class.pe.posting_date,
-                _class.pe.company_gstin,
-                _class.pe_ref.reference_name,
-            )
-
-        query = query.select(*fields)
-        data = query.run(as_dict=True)
-
-        for row in data:
-            self.process_data_for_advances_received_or_adjusted(row, advances_data)
-
-        return advances_data
-
-
-class BooksExcel:
-
-    AMOUNT_HEADERS = [
-        {
-            "fieldname": GSTR1_DataFields.IGST.value,
-            "label": "IGST",
-            "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-        },
-        {
-            "fieldname": GSTR1_DataFields.CGST.value,
-            "label": "CGST",
-            "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-        },
-        {
-            "fieldname": GSTR1_DataFields.SGST.value,
-            "label": "SGST",
-            "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-        },
-        {
-            "fieldname": GSTR1_DataFields.CESS.value,
-            "label": "CESS",
-            "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-        },
-    ]
-
-    def __init__(self, company_gstin, month_or_quarter, year):
-        self.company_gstin = company_gstin
-        self.month_or_quarter = month_or_quarter
-        self.year = year
-
-        self.period = get_period(month_or_quarter, year)
-        doc = frappe.get_doc("GSTR-1 Filed Log", f"{self.period}-{company_gstin}")
-        self.data = doc.load_data("books")["books"]
-
-    def export_data(self):
-        excel = ExcelExporter()
-        excel.remove_sheet("Sheet")
-
-        default_data_format = {"height": 15}
-
-        excel.create_sheet(
-            sheet_name="invoices",
-            headers=self.get_document_headers(),
-            data=self.get_document_data(),
-            default_data_format=default_data_format,
-            add_totals=False,
-        )
-        if hsn_data := self.data.get(GSTR1_SubCategories.HSN.value):
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.HSN.value,
-                headers=self.get_hsn_summary_headers(),
-                data=hsn_data,
-                default_data_format=default_data_format,
-                add_totals=False,
-            )
-
-        if at_received_data := self.data.get(GSTR1_SubCategories.AT.value):
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.AT.value,
-                headers=self.get_at_received_headers(),
-                data=at_received_data,
-                default_data_format=default_data_format,
-                add_totals=False,
-            )
-
-        if at_adjusted_data := self.data.get(GSTR1_SubCategories.TXP.value):
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.TXP.value,
-                headers=self.get_at_adjusted_headers(),
-                data=at_adjusted_data,
-                default_data_format=default_data_format,
-                add_totals=False,
-            )
-
-        if doc_issued_data := self.data.get(GSTR1_SubCategories.DOC_ISSUE.value):
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.DOC_ISSUE.value,
-                headers=self.get_doc_issue_headers(),
-                data=doc_issued_data,
-                default_data_format=default_data_format,
-                add_totals=False,
-            )
-
-        excel.export(get_file_name("books", self.company_gstin, self.period))
-
-    def get_document_data(self):
-        category = [
-            GSTR1_SubCategories.B2B_REGULAR.value,
-            GSTR1_SubCategories.B2B_REVERSE_CHARGE.value,
-            GSTR1_SubCategories.SEZWP.value,
-            GSTR1_SubCategories.SEZWOP.value,
-            GSTR1_SubCategories.DE.value,
-            GSTR1_SubCategories.EXPWP.value,
-            GSTR1_SubCategories.EXPWOP.value,
-            GSTR1_SubCategories.B2CL.value,
-            GSTR1_SubCategories.B2CS.value,
-            GSTR1_SubCategories.NIL_EXEMPT.value,
-            GSTR1_SubCategories.CDNR.value,
-            GSTR1_SubCategories.CDNUR.value,
-        ]
-
-        category_data = []
-        for key, values in self.data.items():
-            if key not in category:
-                continue
-
-            if key in (
-                GSTR1_SubCategories.B2CS.value,
-                GSTR1_SubCategories.NIL_EXEMPT.value,
-            ):
-                category_data.extend(values)
-                continue
-
-            for row in values:
-                dict = row
-                for item in row["items"]:
-                    category_data.extend([{**dict, **item}])
-
-        return category_data
-
-    def get_document_headers(self):
-        return [
-            {
-                "label": "Transaction Type",
-                "fieldname": GSTR1_DataFields.TRANSACTION_TYPE.value,
-            },
-            {
-                "label": "Document Date",
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "header_format": {"width": ExcelWidth.DATE.value},
-            },
-            {
-                "label": "Document Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "label": "Customer GSTIN",
-                "fieldname": GSTR1_DataFields.CUST_GSTIN.value,
-                "header_format": {"width": ExcelWidth.GSTIN.value},
-            },
-            {
-                "label": "Customer Name",
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {
-                "label": "Document Type",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Shipping Bill Number",
-                "fieldname": GSTR1_DataFields.SHIPPING_BILL_NUMBER.value,
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "label": "Shipping Bill Date",
-                "fieldname": GSTR1_DataFields.SHIPPING_BILL_DATE.value,
-                "header_format": {"width": ExcelWidth.DATE.value},
-            },
-            {
-                "label": "Port Code",
-                "fieldname": GSTR1_DataFields.SHIPPING_PORT_CODE.value,
-            },
-            {
-                "label": "Reverse Charge",
-                "fieldname": GSTR1_DataFields.REVERSE_CHARGE.value,
-                "header_format": {"width": ExcelWidth.REVERSE_CHARGE.value},
-            },
-            {
-                "label": "Upload Status",
-                "fieldname": GSTR1_DataFields.UPLOAD_STATUS.value,
-            },
-            {
-                "label": "Place of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {"width": ExcelWidth.POS.value},
-            },
-            *self.AMOUNT_HEADERS,
-            {
-                "label": "Document Value",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-            },
-        ]
-
-    def get_at_received_headers(self):
-        return [
-            {
-                "label": "Advance Date",
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "header_format": {"width": ExcelWidth.DATE.value},
-            },
-            {
-                "label": "Payment Entry Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "label": "Customer",
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {
-                "label": "Place of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {"width": ExcelWidth.POS.value},
-            },
-            *self.AMOUNT_HEADERS,
-            {
-                "label": "Amount Received",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-            },
-        ]
-
-    def get_at_adjusted_headers(self):
-        return [
-            {
-                "label": "Adjustment Date",
-                "fieldname": GSTR1_DataFields.DOC_DATE,
-                "header_format": {"width": ExcelWidth.DATE.value},
-            },
-            {
-                "label": "Adjustment Entry Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER,
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "label": "Customer ",
-                "fieldname": GSTR1_DataFields.CUST_NAME,
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {
-                "label": "Place of Supply",
-                "fieldname": GSTR1_DataFields.POS,
-                "header_format": {"width": ExcelWidth.POS.value},
-            },
-            *self.AMOUNT_HEADERS,
-            {
-                "label": "Amount Adjusted",
-                "fieldname": GSTR1_DataFields.DOC_VALUE,
-            },
-        ]
-
-    def get_hsn_summary_headers(self):
-        return [
-            {
-                "label": "HSN Code",
-                "fieldname": GSTR1_DataFields.HSN_CODE.value,
-            },
-            {
-                "label": "Description",
-                "fieldname": GSTR1_DataFields.DESCRIPTION.value,
-                "header_format": {"width": ExcelWidth.DESCRIPTION.value},
-            },
-            {
-                "label": "UOM",
-                "fieldname": GSTR1_DataFields.UOM.value,
-            },
-            {
-                "label": "Total Quantity",
-                "fieldname": GSTR1_DataFields.QUANTITY.value,
-                "header_format": {"width": ExcelWidth.QUANTITY.value},
-            },
-            *self.AMOUNT_HEADERS,
-        ]
-
-    def get_doc_issue_headers(self):
-        return [
-            {
-                "label": "Document Type",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Sr No From",
-                "fieldname": GSTR1_DataFields.FROM_SR.value,
-            },
-            {
-                "label": "Sr No To",
-                "fieldname": GSTR1_DataFields.TO_SR.value,
-            },
-            {
-                "label": "Total Count",
-                "fieldname": GSTR1_DataFields.TOTAL_COUNT.value,
-                "header_format": {"width": ExcelWidth.INVOICE_COUNT.value},
-            },
-            {
-                "label": "Draft Count",
-                "fieldname": GSTR1_DataFields.DRAFT_COUNT.value,
-                "header_format": {"width": ExcelWidth.INVOICE_COUNT.value},
-            },
-            {
-                "label": "Cancelled Count",
-                "fieldname": GSTR1_DataFields.CANCELLED_COUNT.value,
-                "header_format": {"width": ExcelWidth.INVOICE_COUNT.value},
-            },
-        ]
-
-
-class ReconcileExcel:
-    COLOR_PALLATE = frappe._dict(
-        {
-            "dark_gray": "d9d9d9",
-            "light_gray": "f2f2f2",
-            "dark_pink": "e6b9b8",
-            "light_pink": "f2dcdb",
-            "sky_blue": "c6d9f1",
-            "light_blue": "dce6f2",
-            "green": "d7e4bd",
-            "light_green": "ebf1de",
-        }
-    )
-
-    def __init__(self, company_gstin, month_or_quarter, year):
-        self.company_gstin = company_gstin
-        self.month_or_quarter = month_or_quarter
-        self.year = year
-
-        self.period = get_period(month_or_quarter, year)
-        doc = frappe.get_doc("GSTR-1 Filed Log", f"{self.period}-{company_gstin}")
-
-        self.summary = doc.load_data("reconcile_summary")["reconcile_summary"]
-        self.data = doc.load_data("reconcile")["reconcile"]
-
-    def export_data(self):
-        excel = ExcelExporter()
-        excel.remove_sheet("Sheet")
-
-        default_header_format = {"bg_color": self.COLOR_PALLATE.dark_gray}
-        default_data_format = {"bg_color": self.COLOR_PALLATE.light_gray}
-
-        excel.create_sheet(
-            sheet_name="reconcile summary",
-            headers=self.get_reconcile_summary_headers(),
-            data=self.get_reconcile_summary_data(),
-            default_data_format=default_data_format,
-            default_header_format=default_header_format,
-            add_totals=False,
-        )
-
-        if b2b_data := self.get_b2b_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.B2B.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_b2b_headers(),
-                data=b2b_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if b2cl_data := self.get_b2cl_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.B2CL.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_b2cl_headers(),
-                data=b2cl_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if exports_data := self.get_exports_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.EXP.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_exports_headers(),
-                data=exports_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if b2cs_data := self.get_b2cs_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.B2CS.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_b2cs_headers(),
-                data=b2cs_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if nil_exempt_data := self.get_nil_exempt_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.NIL_EXEMPT.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_nil_exempt_headers(),
-                data=nil_exempt_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if cdnr_data := self.get_cdnr_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.CDNR.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_cdnr_headers(),
-                data=cdnr_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if cdnur_data := self.get_cdnur_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.CDNUR.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_cdnur_headers(),
-                data=cdnur_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if doc_issue_data := self.get_doc_issue_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.DOC_ISSUE.value,
-                merged_headers=self.get_merge_headers_for_doc_issue(),
-                headers=self.get_doc_issue_headers(),
-                data=doc_issue_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if hsn_data := self.get_hsn_summary_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.HSN.value,
-                merged_headers=self.get_merge_headers_for_hsn_summary(),
-                headers=self.get_hsn_summary_headers(),
-                data=hsn_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if at_data := self.get_at_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.AT.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_at_txp_headers(),
-                data=at_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        if txp_data := self.get_txp_data():
-            excel.create_sheet(
-                sheet_name=GSTR1_Excel_Categories.TXP.value,
-                merged_headers=self.get_merge_headers(),
-                headers=self.get_at_txp_headers(),
-                data=txp_data,
-                default_data_format=default_data_format,
-                default_header_format=default_header_format,
-                add_totals=False,
-            )
-
-        excel.export(get_file_name("reconcile", self.company_gstin, self.period))
-
-    def get_merge_headers(self):
-        return frappe._dict(
-            {
-                "Books": [
-                    "books_" + GSTR1_DataFields.POS.value,
-                    "books_" + GSTR1_DataFields.CESS.value,
-                ],
-                "GSTR-1": [
-                    "gstr_1_" + GSTR1_DataFields.POS.value,
-                    "gstr_1_" + GSTR1_DataFields.CESS.value,
-                ],
-            }
-        )
-
-    def get_merge_headers_for_doc_issue(self):
-        return frappe._dict(
-            {
-                "Books": [
-                    "books_" + GSTR1_DataFields.FROM_SR.value,
-                    "books_" + GSTR1_DataFields.CANCELLED_COUNT.value,
-                ],
-                "GSTR-1": [
-                    "gstr_1_" + GSTR1_DataFields.FROM_SR.value,
-                    "gstr_1_" + GSTR1_DataFields.CANCELLED_COUNT.value,
-                ],
-            }
-        )
-
-    def get_merge_headers_for_hsn_summary(self):
-        return frappe._dict(
-            {
-                "Books": [
-                    "books_" + GSTR1_DataFields.UOM.value,
-                    "books_" + GSTR1_DataFields.CESS.value,
-                ],
-                "GSTR-1": [
-                    "gstr_1_" + GSTR1_DataFields.UOM.value,
-                    "gstr_1_" + GSTR1_DataFields.CESS.value,
-                ],
-            }
-        )
-
-    def get_reconcile_summary_headers(self):
-        headers = [
-            {
-                "fieldname": "description",
-                "label": "Description",
-                "header_format": {"width": ExcelWidth.DESCRIPTION.value},
-            },
-            {
-                "fieldname": "total_taxable_value",
-                "label": "Taxable Value",
-                "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-            },
-            {
-                "fieldname": "total_igst_amount",
-                "label": "IGST",
-                "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-            },
-            {
-                "fieldname": "total_cgst_amount",
-                "label": "CGST",
-                "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-            },
-            {
-                "fieldname": "total_sgst_amount",
-                "label": "SGST",
-                "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-            },
-            {
-                "fieldname": "total_cess_amount",
-                "label": "CESS",
-                "data_format": {"number_format": AMOUNT_DATA_FORMAT},
-            },
-        ]
-        return headers
-
-    def get_reconcile_summary_data(self):
-        excel_data = []
-        for row in self.summary:
-            if row["indent"] == 1:
-                continue
-            excel_data.append(row)
-
-        return excel_data
-
-    def get_b2b_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-                "label": "Document Type",
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Document Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Document No",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_GSTIN.value,
-                "label": "Customer GSTIN",
-                "header_format": {"width": ExcelWidth.GSTIN.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_b2b_data(self):
-        b2b_regular = self.data.get(GSTR1_SubCategories.B2B_REGULAR.value, [])
-        b2b_reverse_charge = self.data.get(
-            GSTR1_SubCategories.B2B_REVERSE_CHARGE.value, []
-        )
-        sezwop = self.data.get(GSTR1_SubCategories.SEZWOP.value, [])
-        sezwp = self.data.get(GSTR1_SubCategories.SEZWP.value, [])
-        deemed_export = self.data.get(GSTR1_SubCategories.DE.value, [])
-
-        b2b_data = b2b_regular + b2b_reverse_charge + sezwop + sezwp + deemed_export
-
-        excel_data = []
-
-        for row in b2b_data:
-            row_dict = self.get_row_dict(row)
-
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_b2cl_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-                "label": "Document Type",
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Document Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Document No",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_b2cl_data(self):
-        b2cl_data = self.data.get(GSTR1_SubCategories.B2CL.value, [])
-
-        excel_data = []
-
-        for row in b2cl_data:
-            row_dict = self.get_row_dict(row)
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_exports_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-                "label": "Document Type",
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Document Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Document No",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.SHIPPING_BILL_NUMBER.value,
-                "label": "Shipping Bill Number",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.SHIPPING_BILL_DATE.value,
-                "label": "Shipping Bill Date",
-                "header_format": {"width": ExcelWidth.DATE.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.SHIPPING_PORT_CODE.value,
-                "label": "Shipping Port Code",
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_exports_data(self):
-        expwp = self.data.get(GSTR1_SubCategories.EXPWP.value, [])
-        expwop = self.data.get(GSTR1_SubCategories.EXPWOP.value, [])
-
-        exports_data = expwp + expwop
-
-        excel_data = []
-
-        for row in exports_data:
-            row_dict = self.get_row_dict(row)
-            row_dict.update(
-                {
-                    GSTR1_DataFields.SHIPPING_BILL_NUMBER.value: row.get(
-                        "shipping_bill_number"
-                    ),
-                    GSTR1_DataFields.SHIPPING_BILL_DATE.value: row.get(
-                        "shipping_bill_date"
-                    ),
-                    GSTR1_DataFields.SHIPPING_PORT_CODE.value: row.get(
-                        "shipping_port_code"
-                    ),
-                }
-            )
-
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_b2cs_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-                "label": "Document Type",
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Document Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Document No",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_b2cs_data(self):
-        b2cs_data = self.data.get(GSTR1_SubCategories.B2CS.value, [])
-
-        excel_data = []
-
-        for row in b2cs_data:
-            row_dict = self.get_row_dict(row)
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_nil_exempt_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-                "label": "Document Type",
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Document Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Document No",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_GSTIN.value,
-                "label": "Customer GSTIN",
-                "header_format": {"width": ExcelWidth.GSTIN.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_nil_exempt_data(self):
-        nil_exempt_data = self.data.get(GSTR1_SubCategories.NIL_EXEMPT.value, [])
-
-        excel_data = []
-
-        for row in nil_exempt_data:
-            row_dict = self.get_row_dict(row)
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_cdnr_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-                "label": "Document Type",
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Document Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Document No",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_GSTIN.value,
-                "label": "Customer GSTIN",
-                "header_format": {"width": ExcelWidth.GSTIN.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_cdnr_data(self):
-        cdnr_data = self.data.get(GSTR1_SubCategories.CDNR.value, [])
-
-        excel_data = []
-
-        for row in cdnr_data:
-            row_dict = self.get_row_dict(row)
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_cdnur_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-                "label": "Document Type",
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Document Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Document No",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_GSTIN.value,
-                "label": "Customer GSTIN",
-                "header_format": {"width": ExcelWidth.GSTIN.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "Place of Supply",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "label": "Reverse Charge",
-                "compare_with": "books_" + GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_cdnur_data(self):
-        cdnr_data = self.data.get(GSTR1_SubCategories.CDNUR.value, [])
-
-        excel_data = []
-
-        for row in cdnr_data:
-            row_dict = self.get_row_dict(row)
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_doc_issue_headers(self):
-        headers = [
-            {"fieldname": GSTR1_DataFields.DOC_TYPE.value, "label": "Document Type"},
-            {
-                "fieldname": "match_status",
-                "label": "Match Status",
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.FROM_SR.value,
-                "label": "SR No From",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.FROM_SR.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TO_SR.value,
-                "label": "SR No To",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TO_SR.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TOTAL_COUNT.value,
-                "label": "Total Count",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TOTAL_COUNT.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.INVOICE_COUNT.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CANCELLED_COUNT.value,
-                "label": "Cancelled Count",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CANCELLED_COUNT.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.INVOICE_COUNT.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.FROM_SR.value,
-                "label": "Sr No From",
-                "compare_with": "books_" + GSTR1_DataFields.FROM_SR.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TO_SR.value,
-                "label": "Sr No To",
-                "compare_with": "books_" + GSTR1_DataFields.TO_SR.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TOTAL_COUNT.value,
-                "label": "Total Count",
-                "compare_with": "books_" + GSTR1_DataFields.TOTAL_COUNT.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.INVOICE_COUNT.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CANCELLED_COUNT.value,
-                "label": "Cancelled Count",
-                "compare_with": "books_" + GSTR1_DataFields.CANCELLED_COUNT.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.INVOICE_COUNT.value,
-                },
-            },
-        ]
-
-        return headers
-
-    def get_doc_issue_data(self):
-        doc_issue_data = self.data.get(GSTR1_SubCategories.DOC_ISSUE.value, [])
-
-        excel_data = []
-
-        for row in doc_issue_data:
-            books = row.get("books", {})
-            gstr_1 = row.get("gov", {})
-            row_dict = {
-                GSTR1_DataFields.DOC_TYPE.value: row.get(
-                    GSTR1_DataFields.DOC_TYPE.value
-                ),
-                "match_status": row.get("match_status"),
-                "books_"
-                + GSTR1_DataFields.FROM_SR.value: books.get(
-                    GSTR1_DataFields.FROM_SR.value
-                ),
-                "books_"
-                + GSTR1_DataFields.TO_SR.value: books.get(GSTR1_DataFields.TO_SR.value),
-                "books_"
-                + GSTR1_DataFields.TOTAL_COUNT.value: books.get(
-                    GSTR1_DataFields.TOTAL_COUNT.value
-                ),
-                "books_"
-                + GSTR1_DataFields.CANCELLED_COUNT.value: (
-                    books.get(GSTR1_DataFields.CANCELLED_COUNT.value) or 0
-                )
-                + (books.get(GSTR1_DataFields.DRAFT_COUNT.value) or 0),
-                "gstr_1_"
-                + GSTR1_DataFields.FROM_SR.value: gstr_1.get(
-                    GSTR1_DataFields.FROM_SR.value
-                ),
-                "gstr_1_"
-                + GSTR1_DataFields.TO_SR.value: gstr_1.get(
-                    GSTR1_DataFields.TO_SR.value
-                ),
-                "gstr_1_"
-                + GSTR1_DataFields.TOTAL_COUNT.value: gstr_1.get(
-                    GSTR1_DataFields.TOTAL_COUNT.value
-                ),
-                "gstr_1_"
-                + GSTR1_DataFields.CANCELLED_COUNT.value: (
-                    gstr_1.get(GSTR1_DataFields.CANCELLED_COUNT.value) or 0
-                )
-                + (gstr_1.get(GSTR1_DataFields.DRAFT_COUNT.value) or 0),
-            }
-
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_hsn_summary_headers(self):
-        headers = [
-            {"fieldname": GSTR1_DataFields.HSN_CODE.value, "label": "HSN Code"},
-            {"fieldname": GSTR1_DataFields.DESCRIPTION.value, "label": "Description"},
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.UOM.value,
-                "label": "UQC",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.UOM.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.QUANTITY.value,
-                "label": "Quantity",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.QUANTITY.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.QUANTITY.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST Amount",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST Amount",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST Amount",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS Amount",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.UOM.value,
-                "label": "UQC",
-                "compare_with": "books_" + GSTR1_DataFields.UOM.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.QUANTITY.value,
-                "label": "Quantity",
-                "compare_with": "books_" + GSTR1_DataFields.QUANTITY.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.QUANTITY.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST Amount",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST Amount",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST Amount",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS Amount",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-        return headers
-
-    def get_hsn_summary_data(self):
-        hsn_summary_data = self.data.get(GSTR1_SubCategories.HSN.value, [])
-
-        excel_data = []
-
-        for row in hsn_summary_data:
-            books = row.get("books", {})
-            gstr_1 = row.get("gov", {})
-
-            row_dict = {
-                GSTR1_DataFields.HSN_CODE.value: row.get(
-                    GSTR1_DataFields.HSN_CODE.value
-                ),
-                GSTR1_DataFields.DESCRIPTION.value: row.get(
-                    GSTR1_DataFields.DESCRIPTION.value
-                ),
-                "match_status": row.get("match_status"),
-                "books_"
-                + GSTR1_DataFields.UOM.value: books.get(GSTR1_DataFields.UOM.value),
-                "books_"
-                + GSTR1_DataFields.QUANTITY.value: books.get(
-                    GSTR1_DataFields.QUANTITY.value
-                ),
-                "books_"
-                + GSTR1_DataFields.TAX_RATE.value: books.get(
-                    GSTR1_DataFields.TAX_RATE.value
-                ),
-                "books_"
-                + GSTR1_DataFields.TAXABLE_VALUE.value: books.get(
-                    GSTR1_DataFields.TAXABLE_VALUE.value
-                ),
-                "books_"
-                + GSTR1_DataFields.IGST.value: books.get(GSTR1_DataFields.IGST.value),
-                "books_"
-                + GSTR1_DataFields.CGST.value: books.get(GSTR1_DataFields.CGST.value),
-                "books_"
-                + GSTR1_DataFields.SGST.value: books.get(GSTR1_DataFields.SGST.value),
-                "books_"
-                + GSTR1_DataFields.CESS.value: books.get(GSTR1_DataFields.CESS.value),
-                "gstr_1_"
-                + GSTR1_DataFields.UOM.value: gstr_1.get(GSTR1_DataFields.UOM.value),
-                "gstr_1_"
-                + GSTR1_DataFields.QUANTITY.value: gstr_1.get(
-                    GSTR1_DataFields.QUANTITY.value
-                ),
-                "gstr_1_"
-                + GSTR1_DataFields.TAX_RATE.value: gstr_1.get(
-                    GSTR1_DataFields.TAX_RATE.value
-                ),
-                "gstr_1_"
-                + GSTR1_DataFields.TAXABLE_VALUE.value: gstr_1.get(
-                    GSTR1_DataFields.TAXABLE_VALUE.value
-                ),
-                "gstr_1_"
-                + GSTR1_DataFields.IGST.value: gstr_1.get(GSTR1_DataFields.IGST.value),
-                "gstr_1_"
-                + GSTR1_DataFields.CGST.value: gstr_1.get(GSTR1_DataFields.CGST.value),
-                "gstr_1_"
-                + GSTR1_DataFields.SGST.value: gstr_1.get(GSTR1_DataFields.SGST.value),
-                "gstr_1_"
-                + GSTR1_DataFields.CESS.value: gstr_1.get(GSTR1_DataFields.CESS.value),
-            }
-
-            self.get_taxable_value_difference(row_dict)
-            self.get_tax_difference(row_dict)
-
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_at_txp_headers(self):
-        return [
-            {
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "label": "Advance Date",
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                    "number_format": DATE_FORMAT,
-                },
-            },
-            {
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "label": "Payment Entry Number",
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "label": "Customer Name",
-                "header_format": {"width": ExcelWidth.NAME.value},
-            },
-            {"fieldname": "match_status", "label": "Match Status"},
-            {
-                "fieldname": "taxable_value_difference",
-                "label": "Taxable Value Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "tax_difference",
-                "label": "Tax Difference",
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_pink,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.dark_pink,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.POS.value,
-                "label": "POS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "books_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_green,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.green,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.POS.value,
-                "label": "POS",
-                "compare_with": "books_" + GSTR1_DataFields.POS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAX_RATE.value,
-                "label": "Tax Rate",
-                "compare_with": "books_" + GSTR1_DataFields.TAX_RATE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "label": "Taxable Value",
-                "compare_with": "books_" + GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.IGST.value,
-                "label": "IGST",
-                "compare_with": "books_" + GSTR1_DataFields.IGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CGST.value,
-                "label": "CGST",
-                "compare_with": "books_" + GSTR1_DataFields.CGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.SGST.value,
-                "label": "SGST",
-                "compare_with": "books_" + GSTR1_DataFields.SGST.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-            {
-                "fieldname": "gstr_1_" + GSTR1_DataFields.CESS.value,
-                "label": "CESS",
-                "compare_with": "books_" + GSTR1_DataFields.CESS.value,
-                "data_format": {
-                    "bg_color": self.COLOR_PALLATE.light_blue,
-                    "number_format": AMOUNT_DATA_FORMAT,
-                },
-                "header_format": {
-                    "bg_color": self.COLOR_PALLATE.sky_blue,
-                },
-            },
-        ]
-
-    def get_at_data(self):
-        at_data = self.data.get(GSTR1_SubCategories.AT.value, [])
-
-        excel_data = []
-        for row in at_data:
-            row_dict = self.get_row_dict(row)
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_txp_data(self):
-        txp_adjusted = self.data.get(GSTR1_SubCategories.TXP.value, [])
-
-        excel_data = []
-        for row in txp_adjusted:
-            row_dict = self.get_row_dict(row)
-            excel_data.append(row_dict)
-
-        return excel_data
-
-    def get_row_dict(self, row):
-        books = row.get("books", {})
-        gstr_1 = row.get("gov", {})
-        doc_date = row.get(GSTR1_DataFields.DOC_DATE.value)
-        row_dict = {
-            GSTR1_DataFields.DOC_DATE.value: getdate(doc_date) if doc_date else "",
-            GSTR1_DataFields.DOC_NUMBER.value: row.get(
-                GSTR1_DataFields.DOC_NUMBER.value
-            ),
-            GSTR1_DataFields.CUST_NAME.value: row.get(GSTR1_DataFields.CUST_NAME.value),
-            GSTR1_DataFields.CUST_GSTIN.value: row.get(
-                GSTR1_DataFields.CUST_GSTIN.value
-            ),
-            GSTR1_DataFields.DOC_TYPE.value: row.get(GSTR1_DataFields.DOC_TYPE.value),
-            "match_status": row.get("match_status"),
-            "books_"
-            + GSTR1_DataFields.POS.value: books.get(GSTR1_DataFields.POS.value),
-            "books_"
-            + GSTR1_DataFields.TAX_RATE.value: books.get(
-                GSTR1_DataFields.TAX_RATE.value
-            ),
-            "books_"
-            + GSTR1_DataFields.REVERSE_CHARGE.value: books.get(
-                GSTR1_DataFields.REVERSE_CHARGE.value
-            ),
-            "books_"
-            + GSTR1_DataFields.TAXABLE_VALUE.value: books.get(
-                GSTR1_DataFields.TAXABLE_VALUE.value
-            ),
-            "books_"
-            + GSTR1_DataFields.IGST.value: books.get(GSTR1_DataFields.IGST.value),
-            "books_"
-            + GSTR1_DataFields.CGST.value: books.get(GSTR1_DataFields.CGST.value),
-            "books_"
-            + GSTR1_DataFields.SGST.value: books.get(GSTR1_DataFields.SGST.value),
-            "books_"
-            + GSTR1_DataFields.CESS.value: books.get(GSTR1_DataFields.CESS.value),
-            "gstr_1_"
-            + GSTR1_DataFields.POS.value: gstr_1.get(GSTR1_DataFields.POS.value),
-            "gstr_1_"
-            + GSTR1_DataFields.TAX_RATE.value: gstr_1.get(
-                GSTR1_DataFields.TAX_RATE.value
-            ),
-            "gstr_1_"
-            + GSTR1_DataFields.REVERSE_CHARGE.value: gstr_1.get(
-                GSTR1_DataFields.REVERSE_CHARGE.value
-            ),
-            "gstr_1_"
-            + GSTR1_DataFields.TAXABLE_VALUE.value: gstr_1.get(
-                GSTR1_DataFields.TAXABLE_VALUE.value
-            ),
-            "gstr_1_"
-            + GSTR1_DataFields.IGST.value: gstr_1.get(GSTR1_DataFields.IGST.value),
-            "gstr_1_"
-            + GSTR1_DataFields.CGST.value: gstr_1.get(GSTR1_DataFields.CGST.value),
-            "gstr_1_"
-            + GSTR1_DataFields.SGST.value: gstr_1.get(GSTR1_DataFields.SGST.value),
-            "gstr_1_"
-            + GSTR1_DataFields.CESS.value: gstr_1.get(GSTR1_DataFields.CESS.value),
-        }
-
-        self.get_taxable_value_difference(row_dict)
-        self.get_tax_difference(row_dict)
-
-        return row_dict
-
-    def get_taxable_value_difference(self, row_dict):
-        row_dict["taxable_value_difference"] = (
-            row_dict["books_" + GSTR1_DataFields.TAXABLE_VALUE.value] or 0
-        ) - (row_dict["gstr_1_" + GSTR1_DataFields.TAXABLE_VALUE.value] or 0)
-
-    def get_tax_difference(self, row_dict):
-        row_dict["tax_difference"] = (
-            (row_dict["books_" + GSTR1_DataFields.IGST.value] or 0)
-            - (row_dict["gstr_1_" + GSTR1_DataFields.IGST.value] or 0)
-            + (
-                (row_dict["books_" + GSTR1_DataFields.CGST.value] or 0)
-                - (row_dict["gstr_1_" + GSTR1_DataFields.CGST.value] or 0)
-            )
-            + (
-                (row_dict["books_" + GSTR1_DataFields.SGST.value] or 0)
-                - (row_dict["gstr_1_" + GSTR1_DataFields.SGST.value] or 0)
-            )
-            + (
-                (row_dict["books_" + GSTR1_DataFields.CESS.value] or 0)
-                - (row_dict["gstr_1_" + GSTR1_DataFields.CESS.value] or 0)
-            )
-        )
+        for key, value in row.items():
+            if isinstance(value, (int, float)):
+                aggregated_row[key] += value
+
+    return aggregated_row
