--- india_compliance/gst_india/utils/gstr_1/gstr_1_export.py
@@ -1,726 +1,0 @@
-"""
-Export GSTR-1 data to excel or json
-"""
-
-from datetime import datetime
-
-import frappe
-
-from india_compliance.gst_india.doctype.gstr_1_beta.gstr_1_beta import get_period
-from india_compliance.gst_india.utils.exporter import ExcelExporter
-from india_compliance.gst_india.utils.gstr_1 import (
-    JSON_CATEGORY_EXCEL_CATEGORY_MAPPING,
-    ExcelWidth,
-    GSTR1_DataFields,
-    GSTR1_Gov_Categories,
-    GSTR1_ItemFields,
-    get_file_name,
-)
-from india_compliance.gst_india.utils.gstr_1.gstr_1_json_map import (
-    get_category_wise_data as _get_category_wise_data,
-)
-
-
-class GovExcel:
-    """
-    Export GSTR-1 data to excel
-    """
-
-    AMOUNT_DATA_FORMAT = {
-        "number_format": "#,##0.00",
-    }
-
-    PERCENTAGE_DATA_FORMAT = {
-        "number_format": "0.00",
-    }
-
-    DATE_FORMAT = {
-        "number_format": "dd-mmm-yy",
-    }
-
-    def generate(self, gstin, period):
-        """
-        Build excel file
-        """
-        self.gstin = gstin
-        self.period = period
-        gstr_1_log = frappe.get_doc("GSTR-1 Filed Log", f"{period}-{gstin}")
-
-        self.file_field = "filed" if gstr_1_log.filed else "books"
-        data = gstr_1_log.load_data(self.file_field)[self.file_field]
-        data = process_data(data)
-        self.build_excel(data)
-
-    def build_excel(self, data):
-        excel = ExcelExporter()
-        for category, cat_data in data.items():
-            excel.create_sheet(
-                sheet_name=JSON_CATEGORY_EXCEL_CATEGORY_MAPPING.get(category, category),
-                headers=self.get_category_headers(category),
-                data=cat_data,
-                add_totals=False,
-                default_data_format={"height": 15},
-            )
-
-        excel.remove_sheet("Sheet")
-        excel.export(get_file_name("gov", self.gstin, self.period))
-
-    def get_category_headers(self, category):
-        return getattr(self, f"get_{category.lower()}_headers")()
-
-    def get_b2b_headers(self):
-        return [
-            {
-                "label": "GSTIN/UIN of Recipient",
-                "fieldname": GSTR1_DataFields.CUST_GSTIN.value,
-                "header_format": {
-                    "width": ExcelWidth.GSTIN.value,
-                },
-            },
-            {
-                "label": "Receiver Name",
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "header_format": {
-                    "width": ExcelWidth.NAME.value,
-                },
-            },
-            {
-                "label": "Invoice Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "header_format": {
-                    "width": ExcelWidth.INVOICE_NUMBER.value,
-                },
-            },
-            {
-                "label": "Invoice date",
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "data_format": self.DATE_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                },
-            },
-            {
-                "label": "Invoice Value",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Place Of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "label": "Reverse Charge",
-                "fieldname": GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {"horizontal": "center"},
-                "header_format": {
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "label": "Applicable % of Tax Rate",
-                "fieldname": GSTR1_DataFields.DIFF_PERCENTAGE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DIFF_PERCENTAGE.value,
-                },
-            },
-            {
-                "label": "Invoice Type",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "E-Commerce GSTIN",
-                # Ignore value, just keep the column
-                "fieldname": f"_{GSTR1_DataFields.ECOMMERCE_GSTIN.value}",
-            },
-            {
-                "label": "Taxable Value",
-                "fieldname": GSTR1_ItemFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_ItemFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-        ]
-
-    def get_exp_headers(self):
-        return [
-            {
-                "label": "Export Type",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Invoice Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "header_format": {
-                    "width": ExcelWidth.INVOICE_NUMBER.value,
-                },
-            },
-            {
-                "label": "Invoice date",
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "data_format": self.DATE_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                },
-            },
-            {
-                "label": "Invoice Value",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Port Code",
-                "fieldname": GSTR1_DataFields.SHIPPING_PORT_CODE.value,
-            },
-            {
-                "label": "Shipping Bill Number",
-                "fieldname": GSTR1_DataFields.SHIPPING_BILL_NUMBER.value,
-                "header_format": {
-                    "width": ExcelWidth.INVOICE_NUMBER.value,
-                },
-            },
-            {
-                "label": "Shipping Bill Date",
-                "fieldname": GSTR1_DataFields.SHIPPING_BILL_DATE.value,
-                "data_format": self.DATE_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                },
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Taxable Value",
-                "fieldname": GSTR1_ItemFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_ItemFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-        ]
-
-    def get_b2cl_headers(self):
-        return [
-            {
-                "label": "Invoice Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "header_format": {
-                    "width": ExcelWidth.INVOICE_NUMBER.value,
-                },
-            },
-            {
-                "label": "Invoice date",
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "data_format": self.DATE_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                },
-            },
-            {
-                "label": "Invoice Value",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Place Of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "label": "Applicable % of Tax Rate",
-                "fieldname": GSTR1_DataFields.DIFF_PERCENTAGE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DIFF_PERCENTAGE.value,
-                },
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Taxable Value",
-                "fieldname": GSTR1_ItemFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_ItemFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "E-Commerce GSTIN",
-                # Ignore value, just keep the column
-                "fieldname": f"_{GSTR1_DataFields.ECOMMERCE_GSTIN.value}",
-            },
-        ]
-
-    def get_b2cs_headers(self):
-        return [
-            {
-                "label": "Type",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Place Of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "label": "Applicable % of Tax Rate",
-                "fieldname": GSTR1_DataFields.DIFF_PERCENTAGE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DIFF_PERCENTAGE.value,
-                },
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Taxable Value",
-                "fieldname": GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_DataFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "E-Commerce GSTIN",
-                # Ignore value, just keep the column
-                "fieldname": f"_{GSTR1_DataFields.ECOMMERCE_GSTIN.value}",
-            },
-        ]
-
-    def get_nil_headers(self):
-        return [
-            {
-                "label": "Description",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Nil Rated Supplies",
-                "fieldname": GSTR1_DataFields.NIL_RATED_AMOUNT.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Exempted(other than nil rated/non GST supply)",
-                "fieldname": GSTR1_DataFields.EXEMPTED_AMOUNT.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Non-GST Supplies",
-                "fieldname": GSTR1_DataFields.NON_GST_AMOUNT.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-        ]
-
-    def get_cdnr_headers(self):
-        return [
-            {
-                "label": "GSTIN/UIN of Recipient",
-                "fieldname": GSTR1_DataFields.CUST_GSTIN.value,
-                "header_format": {
-                    "width": ExcelWidth.GSTIN.value,
-                },
-            },
-            {
-                "label": "Receiver Name",
-                "fieldname": GSTR1_DataFields.CUST_NAME.value,
-                "header_format": {
-                    "width": ExcelWidth.NAME.value,
-                },
-            },
-            {
-                "label": "Note Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "header_format": {
-                    "width": ExcelWidth.INVOICE_NUMBER.value,
-                },
-            },
-            {
-                "label": "Note date",
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "data_format": self.DATE_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                },
-            },
-            {
-                "label": "Note Type",
-                "fieldname": GSTR1_DataFields.TRANSACTION_TYPE.value,
-            },
-            {
-                "label": "Place Of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "label": "Reverse Charge",
-                "fieldname": GSTR1_DataFields.REVERSE_CHARGE.value,
-                "data_format": {"horizontal": "center"},
-                "header_format": {
-                    "width": ExcelWidth.REVERSE_CHARGE.value,
-                },
-            },
-            {
-                "label": "Note Supply Type",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Note Value",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Applicable % of Tax Rate",
-                "fieldname": GSTR1_DataFields.DIFF_PERCENTAGE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DIFF_PERCENTAGE.value,
-                },
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Taxable Value",
-                "fieldname": GSTR1_ItemFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_ItemFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-        ]
-
-    def get_cdnur_headers(self):
-        return [
-            {
-                "label": "UR Type",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Note Number",
-                "fieldname": GSTR1_DataFields.DOC_NUMBER.value,
-                "header_format": {
-                    "width": ExcelWidth.INVOICE_NUMBER.value,
-                },
-            },
-            {
-                "label": "Note date",
-                "fieldname": GSTR1_DataFields.DOC_DATE.value,
-                "data_format": self.DATE_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DATE.value,
-                },
-            },
-            {
-                "label": "Note Type",
-                "fieldname": GSTR1_DataFields.TRANSACTION_TYPE.value,
-            },
-            {
-                "label": "Place Of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "label": "Note Value",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Applicable % of Tax Rate",
-                "fieldname": GSTR1_DataFields.DIFF_PERCENTAGE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DIFF_PERCENTAGE.value,
-                },
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Taxable Value",
-                "fieldname": GSTR1_ItemFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_ItemFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-        ]
-
-    def get_at_headers(self):
-        return [
-            {
-                "label": "Place Of Supply",
-                "fieldname": GSTR1_DataFields.POS.value,
-                "header_format": {
-                    "width": ExcelWidth.POS.value,
-                },
-            },
-            {
-                "label": "Applicable % of Tax Rate",
-                "fieldname": GSTR1_DataFields.DIFF_PERCENTAGE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.DIFF_PERCENTAGE.value,
-                },
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Gross Advance Received",
-                "fieldname": GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_DataFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-        ]
-
-    def get_txpd_headers(self):
-        return self.get_at_headers()
-
-    def get_hsn_headers(self):
-        return [
-            {
-                "label": "HSN",
-                "fieldname": GSTR1_DataFields.HSN_CODE.value,
-            },
-            {
-                "label": "Description",
-                "fieldname": GSTR1_DataFields.DESCRIPTION.value,
-            },
-            {
-                "label": "UQC",
-                "fieldname": GSTR1_DataFields.UOM.value,
-            },
-            {
-                "label": "Total Quantity",
-                "fieldname": GSTR1_DataFields.QUANTITY.value,
-                "header_format": {
-                    "width": ExcelWidth.QUANTITY.value,
-                },
-            },
-            {
-                "label": "Total Value",
-                "fieldname": GSTR1_DataFields.DOC_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Rate",
-                "fieldname": GSTR1_DataFields.TAX_RATE.value,
-                "data_format": self.PERCENTAGE_DATA_FORMAT,
-                "header_format": {
-                    "width": ExcelWidth.TAX_RATE.value,
-                },
-            },
-            {
-                "label": "Taxable Value",
-                "fieldname": GSTR1_DataFields.TAXABLE_VALUE.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Integrated Tax Amount",
-                "fieldname": GSTR1_DataFields.IGST.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Central Tax Amount",
-                "fieldname": GSTR1_DataFields.CGST.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "State/UT Tax Amount",
-                "fieldname": GSTR1_DataFields.SGST.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-            {
-                "label": "Cess Amount",
-                "fieldname": GSTR1_DataFields.CESS.value,
-                "data_format": self.AMOUNT_DATA_FORMAT,
-            },
-        ]
-
-    def get_doc_issue_headers(self):
-        return [
-            {
-                "label": "Nature of Document",
-                "fieldname": GSTR1_DataFields.DOC_TYPE.value,
-            },
-            {
-                "label": "Sr. No. From",
-                "fieldname": GSTR1_DataFields.FROM_SR.value,
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "label": "Sr. No. To",
-                "fieldname": GSTR1_DataFields.TO_SR.value,
-                "header_format": {"width": ExcelWidth.INVOICE_NUMBER.value},
-            },
-            {
-                "label": "Total Number",
-                "fieldname": GSTR1_DataFields.TOTAL_COUNT.value,
-                "header_format": {"width": ExcelWidth.INVOICE_COUNT.value},
-            },
-            {
-                "label": "Cancelled",
-                "fieldname": GSTR1_DataFields.CANCELLED_COUNT.value,
-                "header_format": {"width": ExcelWidth.INVOICE_COUNT.value},
-            },
-        ]
-
-
-CATEGORIES_WITH_ITEMS = {
-    GSTR1_Gov_Categories.B2B.value,
-    GSTR1_Gov_Categories.B2CL.value,
-    GSTR1_Gov_Categories.EXP.value,
-    GSTR1_Gov_Categories.CDNR.value,
-    GSTR1_Gov_Categories.CDNUR.value,
-}
-
-
-@frappe.whitelist()
-def export_gstr_1_to_excel(company_gstin, month_or_quarter, year):
-    GovExcel().generate(company_gstin, get_period(month_or_quarter, year))
-
-
-def get_category_wise_data(input_data):
-    return {
-        category: flatten_to_invoice_list(data)
-        for category, data in _get_category_wise_data(input_data).items()
-    }
-
-
-def create_row(invoice_data, item):
-    return {
-        **invoice_data,
-        GSTR1_ItemFields.TAX_RATE.value: item.get(GSTR1_ItemFields.TAX_RATE.value, 0),
-        GSTR1_ItemFields.TAXABLE_VALUE.value: item.get(
-            GSTR1_ItemFields.TAXABLE_VALUE.value, 0
-        ),
-        GSTR1_ItemFields.CESS.value: item.get(GSTR1_ItemFields.CESS.value, 0),
-    }
-
-
-def flatten_to_invoice_list(input_data):
-    return [document for documents in input_data.values() for document in documents]
-
-
-def flatten_invoice_items_to_rows(input_data):
-    return [
-        create_row(invoice, item)
-        for invoice in input_data
-        for item in invoice[GSTR1_DataFields.ITEMS.value]
-    ]
-
-
-FIELD_TRANSFORMATIONS = {
-    GSTR1_DataFields.DIFF_PERCENTAGE.value: lambda value: (
-        value * 100 if value != 0 else None
-    ),
-    GSTR1_DataFields.DOC_DATE.value: lambda value: datetime.strptime(value, "%Y-%m-%d"),
-    GSTR1_DataFields.SHIPPING_BILL_DATE.value: lambda value: datetime.strptime(
-        value, "%Y-%m-%d"
-    ),
-}
-
-
-def modify_row(row):
-    for field, modifier in FIELD_TRANSFORMATIONS.items():
-        if field in row:
-            row[field] = modifier(row[field])
-    return row
-
-
-def process_data(input_data):
-    category_wise_data = get_category_wise_data(input_data)
-
-    processed_data = {
-        category: [
-            modify_row(row)
-            for row in (
-                flatten_invoice_items_to_rows(data)
-                if category in CATEGORIES_WITH_ITEMS
-                else data
-            )
-        ]
-        for category, data in category_wise_data.items()
-    }
-
-    # calculate document_value for HSN
-    for row in processed_data.get(GSTR1_Gov_Categories.HSN.value, []):
-        row[GSTR1_DataFields.DOC_VALUE.value] = sum(
-            (
-                row.get(GSTR1_DataFields.TAXABLE_VALUE.value, 0),
-                row.get(GSTR1_DataFields.IGST.value, 0),
-                row.get(GSTR1_DataFields.CGST.value, 0),
-                row.get(GSTR1_DataFields.SGST.value, 0),
-                row.get(GSTR1_DataFields.CESS.value, 0),
-            )
-        )
-
-    return processed_data
